% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering algorithm}
\usage{
clustering(signatures, k)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}

\item{k}{K value or list of k's to be used as number of clusters}
}
\value{
An extended data frame which includes the k (cluster index) for each signature
}
\description{
Apply the clustering algorithm to group samples into k clusters
}
\seealso{
\code{\link{load_files}}
}
