% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_endmembers}
\alias{clustering_endmembers}
\title{Clustering fixing centroids to endmembers}
\usage{
clustering_endmembers(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be unmixed (using unmixing_fixed or unmixing_vca).}
}
\value{
An extended spectral object which includes the clustering indices of the signatures.
}
\description{
Assign an endmember to each signature based on its distance, similarly
to k-means clustering being fixed the centroids to the endmembers
}
\examples{
\donttest{
data(signatures)
unmixed_signatures <- unmixing_fixed(signatures, files=c("almagre.asd.txt", "blanco.asd.txt"))
clustered_signatures <- clustering_endmembers(unmixed_signatures)
}

}
\seealso{
\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
