% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_kmeans}
\alias{clustering_kmeans}
\title{K-Means clustering}
\usage{
clustering_kmeans(signatures, k)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function}

\item{k}{K value to be used as number of clusters}
}
\value{
An extended spectral object which includes the clustering indices of the signatures
}
\description{
Apply the k-means clustering algorithm to group samples into k clusters
}
\seealso{
\code{\link{load_signature_files}}
}
