% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing.R
\name{endmember_files}
\alias{endmember_files}
\title{Find the files which match with the computed endmembers}
\usage{
endmember_files(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function. set_endmembers must be called before of this function}
}
\value{
A data frame which contains the files which  match with the computed endmembers
}
\description{
Find the files which match with the computed endmembers
}
\seealso{
\code{\link{set_endmembers}}

\code{\link{load_files}}
}
