% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{genReport}
\alias{genReport}
\title{Create a new report}
\usage{
genReport(input_source, path = "/tmp/spectral", title, kclusters, kendmembers)
}
\arguments{
\item{input_source}{A spectral object built using the load_files function or the path of the
folder which contains the .asd.txt files (which include the wavelength measurements)}

\item{path}{The output path where documents are saved. By default, "/tmp/spectral"}

\item{title}{The title included at the begining of the reports}

\item{kclusters}{The number of clusters to group the data}

\item{kendmembers}{The number of endmembers to find in the data}

\item{data_path}{The path for}
}
\value{
It creates a report in html, latex, pdf and word formats
}
\description{
Create a new report
}
\seealso{
\code{\link{load_files}}
}
