% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{is.spectral}
\alias{is.spectral}
\title{Check if the object x is of class spectral}
\usage{
is.spectral(x)
}
\arguments{
\item{x}{Object to check if it is of class spectral}
}
\value{
TRUE if x is of class spectral. False otherwise
}
\description{
Check if the object x is of class spectral
}
