% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_signature_files}
\alias{load_signature_files}
\title{Load the samples from a folder path}
\usage{
load_signature_files(path, header = NULL, wavelength_start = NULL,
  wavelength_end = NULL)
}
\arguments{
\item{path}{Folder path to search for .asd.txt files}

\item{header}{The line format of the head of the data (the line before the data measurements)
If NULL, default to wavelength<tab><filename>}

\item{wavelength_start}{The shorter measured wavelength in nm. If NULL, default to 350}

\item{wavelength_end}{The longer measured wavelength in nm. If NULL, default to 2500}
}
\value{
A spectral object. Contains the information of the signatures, the processed files and the wavelength ranges
}
\description{
It opens every file wich .asd.txt extension and process them:
1) Search for a line with the header format
2) Every line after that is supposed to be <wavelength><tab><value>
}
