% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot the signatures coloured by its cluster asignment.
The function clustering must be run before of this one.}
\usage{
plot_cluster(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
Plot the signatures coloured by its cluster asignment.
The function clustering must be run before of this one.
}
\seealso{
\code{\link{clustering}}

\code{\link{load_files}}
}
