% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clustering}
\usage{
plot_clusters(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be clustered (using clustering_kmeans or clustering_endmembers).}
}
\value{
A ggplot2 object
}
\description{
Plot the signatures coloured by its cluster asignment.
}
\examples{
data(signatures)
clustered_signatures <- clustering_kmeans(signatures, k=3)
plot_clusters(clustered_signatures)

}
\seealso{
\code{\link{clustering_kmeans}}

\code{\link{clustering_endmembers}}

\code{\link{load_signature_files}}
}
