% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Generate the elbow plot}
\usage{
plot_elbow(signatures, k = 1:20, selected = NULL)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}

\item{k}{List of values of k to use in the x-axis}
}
\value{
The plot of the sum of the distances intra-clusters (y-axis) for every number of clusters in k (x-axis)
}
\description{
Plot of the sum of the inter-clusters distances (y-axis) for every number of clusters in k (x-axis)
}
\seealso{
\code{\link{load_files}}
}
