% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmember_cluster}
\alias{plot_endmember_cluster}
\title{Plot the endmembers using the colors of the clusters that they belongs to.}
\usage{
plot_endmember_cluster(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
Plot the endmembers using the colors of the clusters that they belongs to.
}
\seealso{
\code{\link{clustering}}

\code{\link{set_endmembers}}

\code{\link{load_files}}
}
