% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmember_density_bar}
\alias{plot_endmember_density_bar}
\title{Plot the mean composition of each signature expressed as a combination of endmembers, grouped by clusters.}
\usage{
plot_endmember_density_bar(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
The result are a set of bar plots
}
\seealso{
\code{\link{clustering}}

\code{\link{set_endmembers}}

\code{\link{load_files}}
}
