% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmembers}
\alias{plot_endmembers}
\title{Plot the endmember signatures. The function set_endmembers must be run before of this one.}
\usage{
plot_endmembers(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
Plot the endmember signatures. The function set_endmembers must be run before of this one.
}
\seealso{
\code{\link{set_endmembers}}

\code{\link{load_files}}
}
