% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_intracorrelation}
\alias{plot_intracorrelation}
\title{Plot of the intracorrelation among the wavelengths of the signatures.}
\usage{
plot_intracorrelation(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\value{
A ggplot object
}
\description{
Plot of the intracorrelation among the wavelengths of the signatures.
}
\seealso{
\code{\link{load_files}}
}
