% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mutualinfo}
\alias{plot_mutualinfo}
\title{Mutual information plot}
\usage{
plot_mutualinfo(signatures, bins = 10)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function}
}
\value{
A ggplot2 object
}
\description{
Plot of the mutual information among wavelengths for all the signatures.
}
\seealso{
\code{\link{load_signature_files}}
}
