% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plot all the signatures together using one color per sample.}
\usage{
plot_signatures(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\value{
A ggplot object
}
\description{
Plot all the signatures together using one color per sample.
}
\seealso{
\code{\link{load_files}}
}
