% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{print.spectral}
\alias{print.spectral}
\title{Print a short description of the spectral object}
\usage{
\method{print}{spectral}(x, ...)
}
\arguments{
\item{x}{Object of class spectral}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Print a short description of the spectral object
}
