% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{remove_head}
\alias{remove_head}
\title{Filter out the high wavelength values}
\usage{
remove_head(signatures, head)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.}

\item{head}{The wavelengths from the minimum to head will be removed.}
}
\value{
The processed (low wavelength filtered) signatures.
}
\description{
Data preprocessing to remove the lowest wavelengths for each signature, possibly because of noise.
}
\examples{
data(signatures)
filtered_signatures <- remove_head(signatures, head=400)

}
\seealso{
\code{\link{load_signature_files}}
}
