% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing.R
\name{set_endmembers}
\alias{set_endmembers}
\title{VCA algorithm to obtain the k endmembers from the signatures}
\usage{
set_endmembers(signatures, k)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}

\item{k}{K value or list of k's to be used as number of endmembers}
}
\value{
An extended spectral object which includes the indices of the endmember signatures
}
\description{
VCA algorithm to obtain the k endmembers from the signatures
}
\seealso{
\code{\link{load_files}}
}
