% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{summary.spectral}
\alias{summary.spectral}
\title{Print the summary of the spectral object}
\usage{
\method{summary}{spectral}(object, ...)
}
\arguments{
\item{object}{Object of class spectral}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Print the summary of the spectral object
}
