% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_endmembers}
\alias{table_endmembers}
\title{Endmembers description}
\usage{
table_endmembers(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function. The spectral object needs to be unmixed (using unmixing_fixed or unmixing_vca).}
}
\description{
Generate the table which contains the endmembers and the files which they corresponds to.
}
\seealso{
\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
