% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_residuals}
\alias{table_residuals}
\title{Represents each signature as a weighted combination of endmembers.
This function generates a table which contains the residuals of this adjustment: the sum of the errors between the represented and the real values.}
\usage{
table_residuals(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
Represents each signature as a weighted combination of endmembers.
This function generates a table which contains the residuals of this adjustment: the sum of the errors between the represented and the real values.
}
\seealso{
\code{\link{clustering}}

\code{\link{set_endmembers}}

\code{\link{load_files}}
}
