% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmember_cluster}
\alias{plot_endmember_cluster}
\title{Plot endmembers coloured by clustering}
\usage{
plot_endmember_cluster(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be clustered (using clustering_kmeans or clustering_endmembers)
and unmixed (using unmixing_fixed or unmixing_vca).}
}
\value{
A ggplot2 object.
}
\description{
Plot the endmembers using the colors of the clusters that they belongs to.
}
\examples{
data(signatures)
unmixed_signatures <- unmixing_fixed(signatures, files=c("almagre.asd.txt", "blanco.asd.txt"))
clustered_signatures <- clustering_kmeans(unmixed_signatures, k=2)
plot_endmember_cluster(clustered_signatures)

}
\seealso{
\code{\link{clustering_kmeans}}

\code{\link{clustering_endmembers}}

\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
