% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmember_density_bar}
\alias{plot_endmember_density_bar}
\title{Plot of endmember density bar by clusters}
\usage{
plot_endmember_density_bar(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be clustered (using clustering_kmeans or clustering_endmembers)
and unmixed (using unmixing_fixed or unmixing_vca).}
}
\value{
A ggplot2 object
}
\description{
Plot the mean composition of each signature expressed as a combination of endmembers, grouped by clusters.
}
\details{
The result is a set of bar plots
}
\seealso{
\code{\link{clustering_kmeans}}

\code{\link{clustering_endmembers}}

\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
