% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_endmembers}
\alias{plot_endmembers}
\title{Plot endmember signatures}
\usage{
plot_endmembers(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be unmixed (using unmixing_fixed or unmixing_vca).}
}
\value{
A ggplot2 object
}
\description{
Plot the endmember signatures.
}
\examples{
data(signatures)
unmixed_signatures <- unmixing_fixed(signatures, files=c("almagre.asd.txt", "blanco.asd.txt"))
plot_endmembers(unmixed_signatures)

}
\seealso{
\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
