% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_intracorrelation}
\alias{plot_intracorrelation}
\title{Intercorrelation plot}
\usage{
plot_intracorrelation(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.}
}
\value{
A ggplot2 object.
}
\description{
Plot of the intracorrelation among the wavelengths of the signatures.
}
\examples{
data(signatures)
plot_intracorrelation(signatures)

}
\seealso{
\code{\link{load_signature_files}}
}
