% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{smooth_leaps}
\alias{smooth_leaps}
\title{Data preprocessing to smooth the leaps in a number of wavelengths as a consequence of different sensors usage}
\usage{
smooth_leaps(signatures, leaps)
}
\arguments{
\item{leaps}{Wavelengths where exists a leap in the value measurement as a consequence of a different sensor usage}

\item{signature}{A spectral object built using the load_files function}
}
\value{
The processed (smoothed) signatures
}
\description{
Data preprocessing to smooth the leaps in a number of wavelengths as a consequence of different sensors usage
}
\seealso{
\code{\link{load_files}}
}
