% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_residuals_summary}
\alias{table_residuals_summary}
\title{Residuals for each signature}
\usage{
table_residuals_summary(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.
The spectral object needs to be unmixed (using unmixing_fixed or unmixing_vca).}
}
\value{
A matrix object with the tabular information.
}
\description{
Represents each signature as a weighted combination of endmembers.
This function generates a table which contains the residuals of this adjustment: the sum of the errors between the represented and the real values.
}
\examples{
data(signatures)
unmixed_signatures <- unmixing_fixed(signatures, files=c("almagre.asd.txt", "blanco.asd.txt"))
table_residuals_summary(unmixed_signatures)

}
\seealso{
\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
