% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_weights}
\alias{table_weights}
\title{Signature endmember weights}
\usage{
table_weights(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.}
}
\description{
Generate the table which contains the composition of each signature expressed as a weighted combination of endmembers. The spectral object needs to be unmixed (using unmixing_fixed or unmixing_vca).
}
\seealso{
\code{\link{unmixing_fixed}}

\code{\link{unmixing_vca}}

\code{\link{load_signature_files}}
}
