% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_weights}
\alias{table_weights}
\title{Generate the table which contains the composition of each signature expressed as a weighted combination of endmembers.}
\usage{
table_weights(signatures)
}
\arguments{
\item{signatures}{A spectral object built using the load_files function}
}
\description{
Generate the table which contains the composition of each signature expressed as a weighted combination of endmembers.
}
\seealso{
\code{\link{clustering}}

\code{\link{set_endmembers}}

\code{\link{load_files}}
}
