% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing.R
\name{unmixing_fixed}
\alias{unmixing_fixed}
\title{Fix the endmembers}
\usage{
unmixing_fixed(signatures, files, names = NULL, colors = NULL)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function. set_endmembers must be called before of this function}
}
\value{
A data frame which contains the files which  match with the computed endmembers
}
\description{
Find the files which match with the computed endmembers
}
\seealso{
\code{\link{set_endmembers}}

\code{\link{load_signature_files}}
}
