% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmixing.R
\name{unmixing_vca}
\alias{unmixing_vca}
\title{Unmixing based on VCA}
\usage{
unmixing_vca(signatures, k)
}
\arguments{
\item{signatures}{A spectral object built using the load_signature_files function.}

\item{k}{The number of endmembers to find.}
}
\value{
An extended spectral object which includes the indices of the endmember signatures.
}
\description{
VCA algorithm to obtain the k endmembers from the signatures.
}
\examples{
data(signatures)
unmixed_signatures <- unmixing_vca(signatures, k=2)

}
\seealso{
\code{\link{load_signature_files}}
}
